test: adicionando testes unitários para as classes do Problema 2

+++ AnaliseAlgoritmos/Problema_2/IAssinante.cs
@@ -2,5 +2,5 @@
 
 public interface IAssinante
 {
-    void Atualizar(string nomeAcao, decimal novoValor);
+    void Atualizar(string nomeAcao, double novoValor);
 }
